Updated case sensitive & added write file check

+++ src/file_renamer/lib/case.py
@@ -35,6 +35,7 @@ class CaseSensitive(Case):
         logging.info('fr: %s', fr)
 
         self.case_sensitive_val = False
+        self.write = False # File system write permission
         self.new_file_1 = "case-sensitive.txt.tmp"
         self.new_file_2 = "CASE-SENSITIVE.TXT.TMP"
 
@@ -47,33 +48,38 @@ class CaseSensitive(Case):
         count = 0 # Num of files created
         self.new_file_1 = Path(os.path.join(path), self.new_file_1)
         self.new_file_2 = Path(os.path.join(path), self.new_file_2)
-        try:
-            with open(self.new_file_1, 'x') as f1:
-                count += 1
-                logging.info(f"File '{self.new_file_1}' created")
-        except FileExistsError:
-            logging.info(f"File '{self.new_file_1}' already exists")
-        try:
-            with open(self.new_file_2, 'x') as f2:
-                count += 1
-                logging.info(f"File '{self.new_file_2}' created")
-        except FileExistsError:
-            logging.info(f"File '{self.new_file_2}' already exists")
-        logging.info('count: %s', count)
-        if count == 2:
-            self.case_sensitive_val = True
-            logging.info(
-                'self.case_sensitive_val: %s', self.case_sensitive_val
-            )
-            os.remove(self.new_file_1)
-            os.remove(self.new_file_2)
-            logging.info(f"File '{self.new_file_1}' deleted")
-            logging.info(f"File '{self.new_file_2}' deleted")
-        elif count == 1:
-            self.case_sensitive_val = False
-            logging.info(
-                'self.case_sensitive_val: %s', self.case_sensitive_val
-            )
-            os.remove(self.new_file_1)
-            logging.info(f"File '{self.new_file_1}' deleted")
+        if self.fr['write'] is False:
+            try:
+                with open(self.new_file_1, 'x') as f1:
+                    count += 1
+                    self.write = True
+                    self.fr["write"] = self.write
+                    logging.info('self.fr["write"]: %s', self.fr["write"])
+                    logging.info(f"File '{self.new_file_1}' created")
+            except FileExistsError:
+                logging.info(f"File '{self.new_file_1}' already exists")
+            try:
+                with open(self.new_file_2, 'x') as f2:
+                    count += 1
+                    logging.info(f"File '{self.new_file_2}' created")
+            except FileExistsError:
+                logging.info(f"File '{self.new_file_2}' already exists")
+            logging.info('count: %s', count)
+            if count == 2:
+                self.case_sensitive_val = True
+                self.fr['case_sensitive'] = self.case_sensitive_val
+                logging.info(
+                    'self.case_sensitive_val: %s', self.case_sensitive_val
+                )
+                os.remove(self.new_file_1)
+                os.remove(self.new_file_2)
+                logging.info(f"File '{self.new_file_1}' deleted")
+                logging.info(f"File '{self.new_file_2}' deleted")
+            elif count == 1:
+                self.case_sensitive_val = False
+                logging.info(
+                    'self.case_sensitive_val: %s', self.case_sensitive_val
+                )
+                os.remove(self.new_file_1)
+                logging.info(f"File '{self.new_file_1}' deleted")
         return self.case_sensitive_val
